refactor: sdk folder structure
