+++ batcher/aligned-sdk/src/lib.rs
@@ -17,6 +17,5 @@ pub mod eth {
     pub mod batcher_payment_service;
 }
 
-pub mod agg_mode;
 mod beacon;
 pub mod sdk;
